+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.4.1] / 2022-12-15
+### Features
+- ITest - Run all projects with `*.Tests` 
 ### Updated
 - Update `SignExtension` - `throw new PathTooLongException` when length > 260